+++ src/ui/src/views/index/children/full-text-search.vue
@@ -64,40 +64,11 @@
                         </template>
                     </div>
                 </div>
-                <div class="pagination clearfix">
-                    <div class="pagination-info fl">
-                        <span class="mr10">{{$tc("Common['共计N条']", pagination['total'], { N: pagination['total'] })}}</span>
-                        <i18n path="Common['每页显示条数']">
-                            <div ref="paginationSize" place="num"
-                                :class="['pagination-size', { 'active': isShowSizeSetting }]"
-                                :sizeDirection="'auto'"
-                                @click="isShowSizeSetting = !isShowSizeSetting"
-                                v-click-outside="closeSizeSetting">
-                                <span>{{pagination['limit']}}</span>
-                                <i class="bk-icon icon-angle-down"></i>
-                                <transition :name="transformOrigin === 'top' ? 'toggle-slide-top' : 'toggle-slide-bottom'">
-                                    <ul :class="['pagination-size-setting', transformOrigin === 'top' ? 'bottom' : 'top']"
-                                        v-show="sizeTransitionReady">
-                                        <li v-for="(limit, index) in limitList"
-                                            :key="index"
-                                            :class="['size-setting-item', { 'selected': pagination['limit'] === limit }]"
-                                            @click="handleLimitChange(limit)">
-                                            {{limit}}
-                                        </li>
-                                    </ul>
-                                </transition>
-                            </div>
-                        </i18n>
-                    </div>
-                    <bk-paging class="pagination-list fr"
-                        v-if="pagination['count'] > 1"
-                        :type="'compact'"
-                        :total-page="pagination['count']"
-                        :size="'small'"
-                        :cur-page="pagination['current']"
-                        @page-change="handlePageChange">
-                    </bk-paging>
-                </div>
+                <full-text-pagination
+                    :pagination="pagination"
+                    @limit-change="handleLimitChange"
+                    @page-change="handlePageChange">
+                </full-text-pagination>
             </div>
         </div>
         <div class="no-data" v-show="!hasData && showNoData && !searching">
@@ -112,15 +83,16 @@
         addMainScrollListener,
         removeMainScrollListener
     } from '@/utils/main-scroller'
+    import fullTextPagination from './full-text-pagination.vue'
     import { mapGetters, mapActions } from 'vuex'
     export default {
+        components: {
+            fullTextPagination
+        },
         data () {
             return {
                 reg: /(?<=keywords=).*/,
                 scrollTop: null,
-                isShowSizeSetting: false,
-                sizeTransitionReady: false,
-                transformOrigin: 'top',
                 currentClassify: -1,
                 requestId: 'fullTextSearch',
                 query: {
@@ -131,11 +103,9 @@
                 hasData: false,
                 showNoData: false,
                 searching: false,
-                limitList: [10, 20, 50, 100],
                 pagination: {
                     limit: 10,
                     start: 0,
-                    count: 1,
                     current: 1,
                     total: 0
                 },
@@ -212,26 +182,6 @@
                 this.$refs.resultsWrapper.style.marginTop = domRefs.offsetHeight + 'px'
                 addMainScrollListener(this.scrollTop)
             },
-            closeSizeSetting () {
-                this.isShowSizeSetting = false
-            },
-            calcSizePosition () {
-                const paginationSizeDom = this.$refs.paginationSize
-                if (paginationSizeDom.getAttribute('sizeDirection') && paginationSizeDom.getAttribute('sizeDirection') !== 'auto') {
-                    this.transformOrigin = this.pagination.sizeDirection
-                } else {
-                    const sizeSettingItemHeight = 32
-                    const sizeSettingHeight = this.limitList.length * sizeSettingItemHeight
-                    const paginationSizeRect = paginationSizeDom.getBoundingClientRect()
-                    const bodyRect = document.body.getBoundingClientRect()
-                    if (bodyRect.height - paginationSizeRect.y - paginationSizeRect.height > sizeSettingHeight) {
-                        this.transformOrigin = 'bottom'
-                    } else {
-                        this.transformOrigin = 'top'
-                    }
-                }
-                this.sizeTransitionReady = true
-            },
             isPublicModel (objId) {
                 const model = this.models.find(model => model['bk_obj_id'] === objId) || {}
                 return !this.$tools.getMetadataBiz(model)
@@ -290,7 +240,7 @@
                         }
                     }).then(data => {
                         this.searching = false
-                        this.setPagination(data.total)
+                        this.pagination.total = data.total
                         const hitsData = data.hits
                         const modelData = data.aggregations
                         if (hitsData && modelData) {
@@ -389,10 +339,6 @@
                 }
                 const flatternedText = this.$tools.getPropertyText(property, cloneSource)
                 return isHeightLight ? `<em>${flatternedText}</em>` : flatternedText
-            },
-            setPagination (total) {
-                this.pagination.total = total
-                this.pagination.count = Math.ceil(total / this.pagination.limit)
             },
             handleLimitChange (limit) {
                 if (this.pagination.limit === limit) return
@@ -503,83 +449,6 @@
                     }
                 }
             }
-        }
-        .pagination {
-            border-top: 1px solid #dde4eb;
-            padding-top: 10px;
-            color: $cmdbTextColor;
-            font-size: 14px;
-            line-height: 32px;
-            .pagination-size {
-                position: relative;
-                display: inline-block;
-                width: 54px;
-                height: 32px;
-                border: 1px solid #c3cdd7;
-                padding: 0 25px 0 6px;
-                margin-left: 6px;
-                margin-right: 6px;
-                vertical-align: middle;
-                border-radius: 2px;
-                cursor: pointer;
-                .icon-angle-down {
-                    position: absolute;
-                    right: 6px;
-                    top: 9px;
-                    font-size: 12px;
-                    color: #c3cdd7;
-                    transition: transform .2s linear;
-                }
-                &.active {
-                    border-color: #0082ff;
-                    .icon-angle-down {
-                        color: #0082ff;
-                        transform: rotate(180deg);
-                    }
-                }
-            }
-        }
-        .pagination-size-setting{
-            position: absolute;
-            left: 0;
-            width: 100%;
-            background-color: #fff;
-            box-shadow: 0 0 1px 1px rgba(0, 0, 0, 0.1);
-            z-index: 10001;
-            &.top {
-                top: 32px;
-            }
-            &.bottom {
-                bottom: 32px;
-            }
-            .size-setting-item {
-                height: 32px;
-                line-height: 32px;
-                padding: 0 12px;
-                &.selected,
-                &:hover{
-                    background-color: #f5f5f5;
-                    color: #0082ff;
-                }
-            }
-        }
-        .toggle-slide-bottom-enter-active,
-        .toggle-slide-bottom-leave-active,
-        .toggle-slide-top-enter-active,
-        .toggle-slide-top-leave-active {
-            transition: transform .3s cubic-bezier(.23, 1, .23, 1), opacity .5s cubic-bezier(.23, 1, .23, 1);
-            transform-origin: center top;
-        }
-        .toggle-slide-top-enter-active,
-        .toggle-slide-top-leave-active {
-            transform-origin: center bottom;
-        }
-        .toggle-slide-bottom-enter,
-        .toggle-slide-bottom-leave-active,
-        .toggle-slide-top-enter,
-        .toggle-slide-top-leave-active {
-            transform: translateZ(0) scaleY(0);
-            opacity: 0;
         }
         .no-data {
             width: 90%;
